requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
httpx==0.25.2

# AI/ML Dependencies
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
datasets==2.14.6
peft==0.6.2
bitsandbytes==0.41.2
accelerate==0.24.1
huggingface-hub==0.18.0
tokenizers==0.15.0
safetensors==0.4.0

# Vector Database
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.4
scipy==1.11.4

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
konlpy==0.6.0
mecab-python3==1.0.8

# Monitoring & Evaluation
wandb==0.16.0
tensorboard==2.15.1
prometheus-client==0.19.0

# Optimization & Deployment
vllm==0.2.6
triton==2.1.0
onnx==1.15.0
onnxruntime-gpu==1.16.3

# Additional utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
pyyaml==6.0.1
structlog==23.1.0
psutil==5.9.6